// src/payload-types.ts
export interface Media {
  id: number;
  /**
   * Describes the image for screen readers and search engines. Leave empty for videos and purely decorative backgrounds.
   */
  alt?: string | null;
  _objectKey?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}